loose coupling...

+++ src/main/java/com/isoft/customersupport/ticket/TicketService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface TicketService {
 	Ticket createTicket(NewTicketCmd newTicketCmd);
 	Ticket updateTicket(ModifyTicketCmd newTicketCmd);
-	Ticket findTicketById(Long id);
+	Ticket findTicketById(Integer id);
 	Ticket findTicketByCreatorEmail(String email);
 	Ticket findTicketByResolverEmail(String email);
 	List<Ticket> findAllTicket();
-	void deleteTicket(Long id);
+	void deleteTicket(Integer id);
 }
